+++ lib/recommendation/ui/pages/information_page.dart
